+++ tiledb/sm/c_api/tiledb.cc
@@ -1504,6 +1504,25 @@ int32_t tiledb_domain_get_dimension_from_name(
   return TILEDB_OK;
 }
 
+int32_t tiledb_domain_has_dimension(
+    tiledb_ctx_t* ctx,
+    const tiledb_domain_t* domain,
+    const char* name,
+    int32_t* has_dim) {
+  if (sanity_check(ctx) == TILEDB_ERR ||
+      sanity_check(ctx, domain) == TILEDB_ERR) {
+    return TILEDB_ERR;
+  }
+
+  bool b;
+  if (SAVE_ERROR_CATCH(ctx, domain->domain_->has_dimension(name, &b)))
+    return TILEDB_ERR;
+
+  *has_dim = b ? 1 : 0;
+
+  return TILEDB_OK;
+}
+
 /* ****************************** */
 /*           ARRAY SCHEMA         */
 /* ****************************** */
@@ -2031,6 +2050,26 @@ int32_t tiledb_array_schema_get_attribute_from_name(
     save_error(ctx, st);
     return TILEDB_OOM;
   }
+  return TILEDB_OK;
+}
+
+int32_t tiledb_array_schema_has_attribute(
+    tiledb_ctx_t* ctx,
+    const tiledb_array_schema_t* array_schema,
+    const char* name,
+    int32_t* has_attr) {
+  if (sanity_check(ctx) == TILEDB_ERR ||
+      sanity_check(ctx, array_schema) == TILEDB_ERR) {
+    return TILEDB_ERR;
+  }
+
+  bool b;
+  if (SAVE_ERROR_CATCH(
+          ctx, array_schema->array_schema_->has_attribute(name, &b)))
+    return TILEDB_ERR;
+
+  *has_attr = b ? 1 : 0;
+
   return TILEDB_OK;
 }
 
@@ -3139,6 +3178,25 @@ int32_t tiledb_kv_schema_get_attribute_from_name(
     save_error(ctx, st);
     return TILEDB_OOM;
   }
+  return TILEDB_OK;
+}
+
+int32_t tiledb_kv_schema_has_attribute(
+    tiledb_ctx_t* ctx,
+    const tiledb_kv_schema_t* kv_schema,
+    const char* name,
+    int32_t* has_attr) {
+  if (sanity_check(ctx) == TILEDB_ERR ||
+      sanity_check(ctx, kv_schema) == TILEDB_ERR) {
+    return TILEDB_ERR;
+  }
+
+  bool b;
+  if (SAVE_ERROR_CATCH(ctx, kv_schema->array_schema_->has_attribute(name, &b)))
+    return TILEDB_ERR;
+
+  *has_attr = b ? 1 : 0;
+
   return TILEDB_OK;
 }